bot/instagram_bot.py
#!/usr/bin/env python3
"""
Instagram Bot Service Implementation
"""

import os
import logging
import requests
import sys
from supabase import create_client, Client
from instagram_private_api import Client as InstagramClient

class InstagramBot:
    """Instagram Bot Service Class"""

    def __init__(self):
        """Initialize the Instagram Bot"""
        self.logger = logging.getLogger(__name__)

        # Load environment variables
        self.instagram_user = os.getenv("INSTAGRAM_USER", "testuser")
        self.instagram_password = os.getenv("INSTAGRAM_PASSWORD", "testpass")

        # Initialize Supabase client
        self.supabase_url = os.getenv("SUPABASE_URL", "http://localhost:8000")
        self.supabase_key = os.getenv("SUPABASE_KEY", "public-anon-key")
        self.supabase = create_client(self.supabase_url, self.supabase_key)

        self.logger.info("Instagram Bot initialized")

    def run(self):
        """Run the bot's main functionality"""
        self.logger.info("Running bot service")

        # Fetch active triggers from Supabase
        triggers = self._fetch_active_triggers()
        if not triggers:
            self.logger.warning("No active triggers found")
            return

        # Fetch active templates from Supabase
        templates = self._fetch_active_templates()
        if not templates:
            self.logger.warning("No active templates found")
            return

        # Monitor Instagram comments and process them
        self._process_instagram_comments(triggers, templates)

    def _fetch_active_triggers(self):
        """Fetch active triggers from Supabase"""
        self.logger.info("Fetching active triggers from Supabase")
        try:
            response = self.supabase.table('triggers').select('*').execute()
            if response.get('error'):
                self.logger.error(f"Error fetching triggers: {response['error']}")
                return []

            triggers = response.get('data', [])
            active_triggers = [t for t in triggers if t.get('is_active', False)]
            self.logger.info(f"Found {len(active_triggers)} active triggers")
            return active_triggers
        except Exception as e:
            self.logger.error(f"Exception fetching triggers: {e}")
            return []

    def _fetch_active_templates(self):
        """Fetch active templates from Supabase"""
        self.logger.info("Fetching active templates from Supabase")
        try:
            response = self.supabase.table('templates').select('*').execute()
            if response.get('error'):
                self.logger.error(f"Error fetching templates: {response['error']}")
                return []

            templates = response.get('data', [])
            active_templates = [t for t in templates if t.get('is_active', False)]
            self.logger.info(f"Found {len(active_templates)} active templates")
            return active_templates
        except Exception as e:
            self.logger.error(f"Exception fetching templates: {e}")
            return []

    def _process_instagram_comments(self, triggers, templates):
        """Process Instagram comments based on triggers"""
        self.logger.info("Processing Instagram comments")

        # TODO: Implement Instagram API integration
        # For now, we'll simulate comment processing

        # Simulate finding a comment with a trigger keyword
        simulated_comment = {
            "post_id": "12345",
            "user_id": "67890",
            "text": "Hello, I need help!"
        }

        # Check if the comment contains any trigger keywords
        for trigger in triggers:
            keyword = trigger.get('keyword', '').lower()
            if keyword in simulated_comment['text'].lower():
                self.logger.info(f"Trigger keyword '{keyword}' found in comment")

                # Find the appropriate template
                template = next((t for t in templates if t.get('trigger_id') == trigger.get('id')), None)
                if template:
                    self.logger.info(f"Using template: {template.get('content', '')}")

                    # Simulate sending a DM
                    self._send_dm(simulated_comment['user_id'], template)

                    # Log activity
                    self._log_activity(simulated_comment, trigger, template)
                else:
                    self.logger.warning(f"No template found for trigger ID: {trigger.get('id')}")
                break
        else:
            self.logger.info("No trigger keywords found in comment")

    def _send_dm(self, user_id, template):
        """Send a direct message to an Instagram user"""
        self.logger.info(f"Sending DM to user {user_id}")

        # TODO: Implement actual Instagram API DM sending
        # For now, we'll simulate the DM sending

        message = template.get('content', '')
        self.logger.info(f"Message sent: {message}")

    def _log_activity(self, comment, trigger, template):
        """Log bot activity to Supabase"""
        self.logger.info("Logging activity to Supabase")

        activity_data = {
            "comment_id": comment.get("post_id"),
            "user_id": comment.get("user_id"),
            "trigger_id": trigger.get("id"),
            "template_id": template.get("id"),
            "action": "sent_dm",
            "details": f"Responded to '{comment.get('text', '')}' with template '{template.get('content', '')}'"
        }

        try:
            response = self.supabase.table('activity_log').insert(activity_data).execute()
            if response.get('error'):
                self.logger.error(f"Error logging activity: {response['error']}")
            else:
                self.logger.info("Activity logged successfully")
        except Exception as e:
            self.logger.error(f"Exception logging activity: {e}")