bot/requirements.txt
# Bot service dependencies
supabase-py
requests
instagram-private-api